fix: address PR review — share the URL cipher, pin opencv-wasm
Per the automated review on #2671:

- fcd/gen-shipavatar.js duplicated the resource table and cipher functions
  from views/utils/ship-img.ts, with only a comment enforcing parity. If
  the resource table ever changed in ship-img.ts without a matching edit
  here, the generator would silently build wrong URLs and calibrate
  against mismatched art. Extracted the pure cipher logic to
  views/utils/ship-img-cipher.ts (no Electron deps) and have both
  ship-img.ts and gen-shipavatar.js import the same module, the latter
  via babel-hook.js (poi's existing in-process require hook), so the
  script loads the real app module instead of a copy. Verified identical
  output against the old inline copy and re-ran the full generator to
  confirm the URLs/downloads still work.
- opencv-wasm has had no release since 4.3.0-10 (2021); pinned the exact
  version instead of a caret range so a future prerelease bump can't
  silently change the WASM build the generator depends on.
