+++ fcd/gen-shipavatar.js
@@ -39,28 +39,15 @@ const path = require('path')
 const { PNG } = require('pngjs')
 const { parseArgs } = require('util')
 
+// Installs a require() hook that transpiles .ts on the fly with the app's
+// own babel.config.js, so the line below loads the *actual* app module —
+// not a hand-copied reimplementation that could silently drift from it.
+require('../babel-hook')(require('../babel-register.config'))
+const { createCipher } = require('../views/utils/ship-img-cipher')
+
 const CASCADE_URL =
   'https://raw.githubusercontent.com/nagadomi/lbpcascade_animeface/master/lbpcascade_animeface.xml'
 const SHIPAVATAR_PATH = path.resolve(__dirname, 'shipavatar.json')
-
-// Magic of tanaka, kept in sync with views/utils/ship-img.ts
-const resource = [
-  6657, 5699, 3371, 8909, 7719, 6229, 5449, 8561, 2987, 5501, 3127, 9319, 4365, 9811, 9927, 2423,
-  3439, 1865, 5925, 4409, 5509, 1517, 9695, 9255, 5325, 3691, 5519, 6949, 5607, 9539, 4133, 7795,
-  5465, 2659, 6381, 6875, 4019, 9195, 5645, 2887, 1213, 1815, 8671, 3015, 3147, 2991, 7977, 7045,
-  1619, 7909, 4451, 6573, 4545, 8251, 5983, 2849, 7249, 7449, 9477, 5963, 2711, 9019, 7375, 2201,
-  5631, 4893, 7653, 3719, 8819, 5839, 1853, 9843, 9119, 7023, 5681, 2345, 9873, 6349, 9315, 3795,
-  9737, 4633, 4173, 7549, 7171, 6147, 4723, 5039, 2723, 7815, 6201, 5999, 5339, 4431, 2911, 4435,
-  3611, 4423, 9517, 3243,
-]
-
-const createKey = (seed) => [...seed].reduce((sum, c) => sum + c.charCodeAt(0), 0)
-
-const createCipher = (id, seed) =>
-  (
-    ((17 * (id + 7) * resource[(createKey(seed) + id * seed.length) % 100]) % 8973) +
-    1000
-  ).toString()
 
 const shipImgUrl = (host, kind, id, damaged, version) => {
   const ntype = kind + (damaged ? '_dmg' : '')
