+++ package.json
@@ -214,7 +214,7 @@
     "lint-staged": "^16.4.0",
     "node-7z": "^3.0.0",
     "node-unzip-2": "^0.2.8",
-    "opencv-wasm": "^4.3.0-10",
+    "opencv-wasm": "4.3.0-10",
     "pngjs": "^7.0.0",
     "postcss-jsx": "^0.36.4",
     "postcss-styled-syntax": "^0.7.1",
