+++ views/utils/ship-img-cipher.ts
@@ -1,0 +1,34 @@
+// Magic of tanaka — the resource table used to derive kcs2 asset URLs (ship
+// banner/remodel/card, equip icons, etc.). This is the single source of
+// truth: views/utils/ship-img.ts uses it to build in-app asset URLs, and
+// fcd/gen-shipavatar.js (a standalone Node script outside the app's build)
+// loads this same file via babel-hook.js so its downloader can never drift
+// from what the game client actually serves.
+export const resource = [
+  6657, 5699, 3371, 8909, 7719, 6229, 5449, 8561, 2987, 5501, 3127, 9319, 4365, 9811, 9927, 2423,
+  3439, 1865, 5925, 4409, 5509, 1517, 9695, 9255, 5325, 3691, 5519, 6949, 5607, 9539, 4133, 7795,
+  5465, 2659, 6381, 6875, 4019, 9195, 5645, 2887, 1213, 1815, 8671, 3015, 3147, 2991, 7977, 7045,
+  1619, 7909, 4451, 6573, 4545, 8251, 5983, 2849, 7249, 7449, 9477, 5963, 2711, 9019, 7375, 2201,
+  5631, 4893, 7653, 3719, 8819, 5839, 1853, 9843, 9119, 7023, 5681, 2345, 9873, 6349, 9315, 3795,
+  9737, 4633, 4173, 7549, 7171, 6147, 4723, 5039, 2723, 7815, 6201, 5999, 5339, 4431, 2911, 4435,
+  3611, 4423, 9517, 3243,
+]
+
+export function createKey(t: string): number {
+  let e = 0
+  if (null != t && '' != t) {
+    for (let i = 0; i < t.length; i++) {
+      e += t.charCodeAt(i)
+    }
+  }
+  return e
+}
+
+export function createCipher(id: number | string, seed: string): string {
+  const o = id.toString().match(/\d+/)
+  if (null == o || 0 == o.length) return ''
+  const r = parseInt(o[0])
+  const s = createKey(seed)
+  const a = null == seed || 0 == seed.length ? 1 : seed.length
+  return (((17 * (r + 7) * resource[(s + r * a) % 100]) % 8973) + 1e3).toString()
+}
